Changed dir for findtest and added empty result test
